feat: lower stringByteLength cap to 20
