+++ src/utils/web_byte_utils.ts
@@ -1,4 +1,5 @@
 import { BSONError } from '../error';
+import { tryLatin } from './latin';
 
 type TextDecoder = {
   readonly encoding: string;
@@ -173,31 +174,9 @@ export const webByteUtils = {
   },
 
   toUTF8(uint8array: Uint8Array, start: number, end: number, fatal: boolean): string {
-    if (uint8array.length === 0) {
-      return '';
-    }
-
-    const stringByteLength = end - start;
-    if (stringByteLength === 0) {
-      return '';
-    }
-
-    if (stringByteLength < 200) {
-      let basicLatin = true;
-      const latinBytes = [];
-      for (let i = start; i < end; i++) {
-        const byte = uint8array[i];
-        if (byte > 127) {
-          basicLatin = false;
-          break;
-        }
-        latinBytes.push(byte);
-      }
-
-      if (basicLatin) {
-        // eslint-disable-next-line prefer-spread
-        return String.fromCharCode.apply(String, latinBytes);
-      }
+    const basicLatin = end - start <= 20 ? tryLatin(uint8array, start, end) : null;
+    if (basicLatin != null) {
+      return basicLatin;
     }
 
     if (fatal) {
